Update migration files
