+++ orif/user/Database/Seeds/AddUserDatas.php
@@ -10,9 +10,9 @@ class AddUserDatas extends \CodeIgniter\Database\Seeder
     {
         $data = [
             ['fk_user_type'=>1,'username'=>'admin','password'=>'$2y$10$84r63xo.M4LVcIi8IvT8cO0qYxyglPshY1jJmKLedRMcaTcxhcVYO'],
-            ['fk_user_type'=>2,'username'=>'utilisateur','password'=>'$2y$10$11wIuR3FnfWwTpfyJ9WCz.E3KErvb.i.Q2Wef6XMUZHTXUlW0FhJm']
+            ['fk_user_type'=>2,'username'=>'trainer','password'=>'$2y$10$Y5mWEJlmsTHdDNrR8OsqAO12PvH4t/Mc.pNMFhTHFlUDvfnhq4dni','email'=>'trainer@orif.ch'],
+            ['fk_user_type'=>3,'username'=>'apprentice','password'=>'$2y$10$n5MJQ00.TohRsgZmMulkHOZtcnszsQuC14S.sIMpCXT0J8ue0L3za','email'=>'apprentice@orif.ch']
         ];
-        foreach($data as $row)
-        $this->db->table('user')->insert($row);
+        foreach($data as $row) $this->db->table('user')->insert($row);
     }
 }
