+++ orif/user/Database/Seeds/AddUserTypeDatas.php
@@ -4,17 +4,32 @@
 namespace User\Database\Seeds;
 
 
-class AddUserTypeDatas extends \CodeIgniter\Database\Seeder
+use CodeIgniter\Database\Seeder;
+
+class addUserTypeDatas extends Seeder
 {
     public function run()
     {
-        $data=[
-            ['name'=>'Administrateur','access_level'=>4],
-            ['name'=>'Enregistré','access_level'=>2],
-            ['name'=>'Invité','access_level'=>1]
+        //user_type//
+        $user_types=[
+            array (
+                'id' => '1',
+                'name' => 'Administrateur',
+                'access_level' => '4',
+            ),
+            array (
+                'id' => '2',
+                'name' => 'Formateur',
+                'access_level' => '2',
+            ),
+            array (
+                'id' => '3',
+                'name' => 'Apprenti',
+                'access_level' => '1',
+            ),
         ];
-        foreach($data as $row)
-            $this->db->table('user_type')->insert($row);
-
+        foreach ($user_types as $user_type){
+            $this->db->table('user_type')->insert($user_type);
+        }
     }
 }
